too many requests problem fixed, the editor.js file's structure has been changed

+++ db/migrate/20240603171508_remove_url_and_dataset_id_from_saved_datasets.rb
@@ -1,6 +1,13 @@
-class RemoveUrlAndDatasetIdFromSavedDatasets < ActiveRecord::Migration[6.0]
+class RemoveUrlAndDatasetIdFromSavedDatasets < ActiveRecord::Migration[6.1]
   def change
-    remove_column :saved_datasets, :url, :string
-    remove_column :saved_datasets, :dataset_id, :integer
+    # Check if the column :url exists and then remove it
+    if column_exists?(:saved_datasets, :url)
+      remove_column :saved_datasets, :url, :string
+    end
+
+    # Check if the column :dataset_id exists and then remove it
+    if column_exists?(:saved_datasets, :dataset_id)
+      remove_column :saved_datasets, :dataset_id, :integer
+    end
   end
 end
